Complete Minimum Viable Product

+++ app/src/main/java/com/example/adobe_exp_batch_status_android/Objects/Batch.kt
@@ -1,3 +1,3 @@
 package com.example.adobe_exp_batch_status_android.Objects
 
-class Batch(var id: String, var lastUpdated: Int, var status: String) {}
+class Batch(var id: String, var lastUpdated: Long, var status: String) {}
